Move Firebase config/init to sep file
