+++ src/index.tsx
@@ -1,25 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import './firebase'; // Keep as one of the first imports since the app is initialized here
 import './css/index.css';
 import Router from './Router';
 import reportWebVitals from './reportWebVitals';
-import { initializeApp } from 'firebase/app';
-import { getAnalytics } from 'firebase/analytics';
-
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const analytics = getAnalytics(app);
-analytics.app.automaticDataCollectionEnabled = false;
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
